requirements.txt
Pillow==7.2.0
numpy==1.19.0
albumentations==1.1.0
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.48.2